+++ src/Graph.Traversals.Tests/DepthFirstPreOrderTraversalTests.cs
@@ -1,0 +1,13 @@
+﻿using Graph.Indexes;
+
+namespace Graph.Traversals.Tests
+{
+    public sealed class DepthFirstPreOrderTraversalTests
+        : TraversalTests
+    {
+        protected override ITraversal<int> CreateTraversal(IAdjacencyIndex<int> index)
+        {
+            return new DepthFirstPreOrderTraversal<int>(index);
+        }
+    }
+}
