+++ src/Graph.Traversals.Tests/FastDepthFirstPreOrderTraversalTests.cs
@@ -1,0 +1,13 @@
+﻿using Graph.Indexes;
+
+namespace Graph.Traversals.Tests
+{
+    public sealed class FastDepthFirstPreOrderTraversalTests
+        : TraversalTests
+    {
+        protected override ITraversal<int> CreateTraversal(IAdjacencyIndex<int> index)
+        {
+            return new FastDepthFirstPreOrderTraversal(index);
+        }
+    }
+}
